Add get all tours and create new tour

+++ sql/schema/001_tour.sql
@@ -3,15 +3,15 @@ CREATE EXTENSION IF NOT EXISTS "uuid-ossp";
 
 CREATE TABLE tours (
   id UUID PRIMARY KEY DEFAULT uuid_generate_v1(),
-  name VARCHAR(100) NOT NULL UNIQUE ,
-  slug VARCHAR(255) , 
-  duration DECIMAL(3, 1),
+  name VARCHAR(100) NOT NULL UNIQUE,
+  slug VARCHAR(255), 
+  duration DOUBLE PRECISION,
   maxGroupSize INTEGER NOT NULL,
-  difficulty VARCHAR(255) NOT NULL CHECK (difficulty IN ('easy', 'medium', 'difficult')),  
-  ratingAverage DECIMAL(3, 1) DEFAULT 4.5 CHECK (ratingAverage >= 1),
+  difficulty VARCHAR(255) NOT NULL CHECK (difficulty IN ('easy', 'medium', 'difficult')),
+  ratingAverage DOUBLE PRECISION DEFAULT 4.5 CHECK (ratingAverage >= 1),
   ratingQuantity INTEGER DEFAULT 0,
   price INTEGER NOT NULL,
-  discount DECIMAL(5, 1),
+  discount DOUBLE PRECISION,
   summary TEXT NOT NULL,
   description TEXT,
   imageCover TEXT,
@@ -19,6 +19,7 @@ CREATE TABLE tours (
   startDates TIMESTAMP[],
   secretTour BOOLEAN DEFAULT FALSE
 );
+
 
 -- +goose Down
 DROP TABLE tours;
